+++ arcgis_map_platform_interface/lib/src/types/arcgis_map_options.dart
@@ -79,10 +79,10 @@ extension LatLngJsonExtension on LatLng {
 ///
 /// https://developers.arcgis.com/javascript/latest/api-reference/esri-views-View.html#padding
 class ViewPadding {
-  final int left;
-  final int top;
-  final int right;
-  final int bottom;
+  final double left;
+  final double top;
+  final double right;
+  final double bottom;
 
   const ViewPadding({
     this.left = 0,
